+++ frontend/src/pages/Scoreboard/Team.tsx
@@ -1,6 +1,6 @@
 import { Box } from "@mui/material";
 import { useParams } from "react-router-dom";
-import TeamScoreboard from "../components/TeamScoreboard";
+import TeamScoreboard from "../../components/TeamScoreboard";
 
 export default function TeamScoreboardPage() {
   const { team: team } = useParams() as unknown as { team: number };
